+++ .php-cs-fixer.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+use PhpCsFixer\Config;
+use PhpCsFixer\Finder;
+
+$finder = Finder::create()
+    ->in('src')
+    ->in('tests');
+
+$config = new Config();
+
+return $config
+    ->setRiskyAllowed(true)
+    ->setRules([
+        '@PSR12' => true,
+        '@Symfony' => true,
+        'concat_space' => ['spacing' => 'one'],
+        'increment_style' => ['style' => 'post'],
+        'no_superfluous_phpdoc_tags' => true,
+        'phpdoc_summary' => false,
+        'yoda_style' => ['equal' => false, 'identical' => false, 'less_and_greater' => false],
+    ])
+    ->setFinder($finder);